requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.22
mysql-connector-python==8.0.26
python-dotenv==0.19.1
pycryptodome==3.10.1
